Homework 5 -Task_1 and Pizza:

+++ src/homework_5/Alina_Mkhoyan/pizza/CheckPrinter.java
@@ -3,12 +3,12 @@ package homework_5.Alina_Mkhoyan.pizza;
 
 public class CheckPrinter {
     public static void printCheck(Order order) {
-        float totalAmount = 0;
+        double totalAmount = 0;
         System.out.print("*********************************" + "\n" + "Order Time: " + order.getTime()
                 + "\n" + "Order: " + Order.getOrderId() + "\n" + "Client: " +
                 Order.getCustomerId() + "\n");
         for (int i = 0; i < order.getItems().size(); i++) {
-            float amount = 0;
+            double amount = 0;
             System.out.println("Name: " + order.getItems().get(i).getName() + "\n" + "---------------------" + "\n" +
                     "Pizza Base (Calzone)" + " " + order.getItems().get(i).getType().getPrice() + " $");
             for (int j = 0; j < order.getItems().get(i).getIngredients().size(); j++) {
