+++ src/homework_5/Alina_Mkhoyan/pizza/PizzaType.java
@@ -2,22 +2,22 @@ package homework_5.Alina_Mkhoyan.pizza;
 
 
 public enum PizzaType {
-    Regular("regular", 1.0F),
-    Closed("Closed (Calzone)", 1.5F);
+    Regular("regular", 1.0),
+    Closed("Closed (Calzone)", 1.5);
 
-    PizzaType(String name, float price) {
+    PizzaType(String name, double price) {
         this.name = name;
         this.price = price;
     }
 
     private final String name;
-    private final float price;
+    private final double price;
 
     public String getName() {
         return name;
     }
 
-    public float getPrice() {
+    public double getPrice() {
         return price;
     }
 }
